pyproject.toml
[tool.poetry]
name = "helios_package"
version = "0.2.38"
description = "HELIOS"
authors = ["heliosgrounder"]
readme = "README.md"

include = [
  { path = ".pre-commit-config.yaml", format = ["sdist", "wheel"] },
  { path = "setup.cfg", format = ["sdist", "wheel"] },
  { path = "hooks", format = ["sdist", "wheel"] },
  { path = ".git", format = ["sdist", "wheel"] }
]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
pillow = "10.2.0"
numpy = "1.24.3"
opencv-python = "4.7.0.72"
tensorflow-io-gcs-filesystem = "0.31.0"
tensorflow = "2.13.0"
black = "24.3.0"
isort = "5.13.2"
flake8 = "7.0.0"
pre-commit = "3.5.0"
pytest = "8.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.github
  | _build
  | buck-out
  | build
  | dist
  | hooks
)/
'''

[tool.isort]
profile="black"
force_sort_within_sections = "False"
order_by_type = "False"
